% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{shade}
\alias{shade}
\alias{tint}
\title{Tint and shade your colors.}
\usage{
shade(color, steps = 0.33)

tint(color, steps = 0.33)
}
\arguments{
\item{color}{(character) Hex or R color.}

\item{steps}{(double) [0, 1] How far from your color [0] to black [1] or white [1].}
}
\description{
Tint and shade your colors.
}
